+++ packages/themes/src/betaddaDark/betaddaDark.ts
@@ -1,4 +1,5 @@
 import avertaStd from '../sharedTokens/avertaStd';
+import componentsTokens from '../sharedTokens/componentsTokens';
 import sharedTokens, { Theme } from '../sharedTokens/sharedTokens';
 import supportColors from '../supportColors/supportColors';
 import supportiveColors from '../v2/colors/supportiveColors';
@@ -60,7 +61,7 @@ const colorNew = {
 };
 
 const betaddaDark: Theme = {
-  ...sharedTokens(colorNew),
+  ...sharedTokens,
   brand: 'Betadda',
   fontFace: avertaStd.fontFace,
   fontFamily: avertaStd.fontStack,
@@ -79,6 +80,7 @@ const betaddaDark: Theme = {
     ...colorNew,
     ...supportiveColors,
   },
+  ...componentsTokens(colorNew),
   colorScheme: 'dark',
 };
