+++ packages/themes/src/bombayClub/bombayClub.ts
@@ -1,4 +1,5 @@
 import avertaStd from '../sharedTokens/avertaStd';
+import componentsTokens from '../sharedTokens/componentsTokens';
 import sharedTokens, { Theme } from '../sharedTokens/sharedTokens';
 import supportColors from '../supportColors/supportColors';
 import supportiveColors from '../v2/colors/supportiveColors';
@@ -53,7 +54,7 @@ const colorNew = {
 };
 
 const bombayClub: Theme = {
-  ...sharedTokens(colorNew),
+  ...sharedTokens,
   radius: {
     small: 4,
     default: 8,
@@ -75,6 +76,7 @@ const bombayClub: Theme = {
     ...colorNew,
     ...supportiveColors,
   },
+  ...componentsTokens(colorNew),
   colorScheme: 'dark',
 };
