+++ packages/themes/src/commsDark/commsDark.ts
@@ -1,4 +1,5 @@
 import avertaStd from '../sharedTokens/avertaStd';
+import componentsTokens from '../sharedTokens/componentsTokens';
 import sharedTokens, { Theme } from '../sharedTokens/sharedTokens';
 import supportColors from '../supportColors/supportColors';
 import supportiveColors from '../v2/colors/supportiveColors';
@@ -53,7 +54,7 @@ const colorNew = {
 };
 
 const commsDark: Theme = {
-  ...sharedTokens(colorNew),
+  ...sharedTokens,
   brand: 'Comms',
   fontFace: avertaStd.fontFace,
   fontFamily: avertaStd.fontStack,
@@ -77,6 +78,7 @@ const commsDark: Theme = {
     default: 16,
     largest: 24,
   },
+  ...componentsTokens(colorNew),
   colorScheme: 'dark',
 };
