+++ packages/themes/src/luckyslots/luckyslots.ts
@@ -1,4 +1,5 @@
 import avertaStd from '../sharedTokens/avertaStd';
+import componentsTokens from '../sharedTokens/componentsTokens';
 import sharedTokens, { Theme } from '../sharedTokens/sharedTokens';
 import supportColors from '../supportColors/supportColors';
 import supportiveColors from '../v2/colors/supportiveColors';
@@ -58,7 +59,7 @@ const colorNew = {
 };
 
 const luckyslots: Theme = {
-  ...sharedTokens(colorNew),
+  ...sharedTokens,
   brand: 'LuckySlots',
   fontFace: avertaStd.fontFace,
   fontFamily: avertaStd.fontStack,
@@ -76,6 +77,7 @@ const luckyslots: Theme = {
     ...colorNew,
     ...supportiveColors,
   },
+  ...componentsTokens(colorNew),
   colorScheme: 'dark',
 };
