+++ packages/themes/src/shangrilaDark/shangrilaDark.ts
@@ -1,4 +1,5 @@
 import avertaStd from '../sharedTokens/avertaStd';
+import componentsTokens from '../sharedTokens/componentsTokens';
 import sharedTokens, { Theme } from '../sharedTokens/sharedTokens';
 import supportColors from '../supportColors/supportColors';
 import supportiveColors from '../v2/colors/supportiveColors';
@@ -53,7 +54,7 @@ const colorNew = {
 };
 
 const shangrilaDark: Theme = {
-  ...sharedTokens(colorNew),
+  ...sharedTokens,
   brand: 'ShangriLa.io',
   fontFace: avertaStd.fontFace,
   fontFamily: avertaStd.fontStack,
@@ -71,6 +72,7 @@ const shangrilaDark: Theme = {
     ...colorNew,
     ...supportiveColors,
   },
+  ...componentsTokens(colorNew),
   colorScheme: 'dark',
 };
